Correction in circle mesh

+++ SoyuzEngine/Meshes/Line.cs
@@ -18,7 +18,7 @@ namespace Soyuz.Meshes
 
             for(int i = 0; i < PointNumber; i++)
             {
-                double x = 2 * Math.PI * ((i + 1) / PointNumber);
+                double x = 2 * Math.PI * ((i + 1) / (double) PointNumber);
                 msh.Positions.Add(new Vector3( (float) Math.Cos(x), 0, (float) Math.Sin(x)));
                 msh.Normals.Add(new Vector3(0, 1, 0));
                 msh.UVs.Add(new Vector2( (float)Math.Cos(x), (float)Math.Sin(x) ));
